Conflicts:
	RobotCode/MecanumDrive/src/org/usfirst/frc4915/MecanumDrive/Robot.java

+++ RobotCode/MecanumDrive/src/org/usfirst/frc4915/MecanumDrive/commands/elevator/ElevatorJumpToPosition.java
@@ -11,9 +11,9 @@ public class ElevatorJumpToPosition extends Command {
 
 	// Position number -- number of totes that you would need to stack on top
 	// of.
-	private int positionNumber;
+	private double positionNumber;
 
-	public ElevatorJumpToPosition(int position) {
+	public ElevatorJumpToPosition(double position) {
 		// Use requires() here to declare subsystem dependencies
 		// eg. requires(chassis);
 		positionNumber = position;
